+++ devel/xdhdws/xdhdws.h
@@ -206,22 +206,26 @@ namespace xdhdws {
 			C_().Process( xdhcmn::fInsertAfter, NULL, Sibling.Internal()(), Id.Internal()() );
 		}
 		void InsertCSSRule(
+			const rNString &Id,
 			const rNString &Rule,
-			const rNString &Index )
+			const rNString &Index)
 		{
-			C_().Process( xdhcmn::fInsertCSSRule, NULL, Rule.Internal()(), Index.Internal()() );
+			C_().Process( xdhcmn::fInsertCSSRule, NULL, Id.Internal()(), Rule.Internal()(), Index.Internal()() );
 		}
 		const char *AppendCSSRule(
+			const rNString &Id,
 			const rNString &Rule,
 			qCBUFFERr &Buffer )
 		{
-			C_().Process( xdhcmn::fAppendCSSRule, &Buffer, Rule.Internal()() );
+			C_().Process( xdhcmn::fAppendCSSRule, &Buffer, Id.Internal()(), Rule.Internal()() );
 
 			return Buffer;
 		}
-		void RemoveCSSRule(const rNString &Index)
+		void RemoveCSSRule(
+			const rNString &Id,
+			const rNString &Index)
 		{
-			C_().Process( xdhcmn::fRemoveCSSRule, NULL, Index.Internal()() );
+			C_().Process( xdhcmn::fRemoveCSSRule, NULL, Id.Internal()(), Index.Internal()() );
 		}
 		void AddClasses(
 			const rNString &Ids,